+++ bin/proving-service/src/commands/proxy.rs
@@ -9,13 +9,14 @@ use pingora::{
 use pingora_proxy::http_proxy_service;
 use tracing::warn;
 
+use super::ProxyConfig;
 use crate::{
-    error::TxProverServiceError,
+    error::ProvingServiceError,
     proxy::{update_workers::LoadBalancerUpdateService, LoadBalancer, LoadBalancerState},
     utils::MIDEN_PROVING_SERVICE,
 };
 
-/// Starts the proxy defined in the config file.
+/// Starts the proxy.
 ///
 /// Example: `miden-proving-service start-proxy 0.0.0.0:8080 127.0.0.1:9090`
 #[derive(Debug, Parser)]
@@ -25,17 +26,20 @@ pub struct StartProxy {
     /// Example: `127.0.0.1:8080 192.168.1.1:9090`
     #[clap(value_name = "WORKERS")]
     workers: Vec<String>,
+    /// Proxy configurations.
+    #[clap(flatten)]
+    proxy_config: ProxyConfig,
 }
 
 impl StartProxy {
-    /// Starts the proxy defined in the config file.
+    /// Starts the proxy using the configuration defined in the command.
     ///
-    /// This method will first read the config file to get the parameters for the proxy. It will
-    /// then start a proxy with each worker passed as command argument as a backend.
+    /// This method will start a proxy with each worker passed as command argument as a backend,
+    /// using the configurations passed as options for the commands or the equivalent environmental
+    /// variables.
     ///
     /// # Errors
     /// Returns an error in the following cases:
-    /// - The config file cannot be read.
     /// - The backend cannot be created.
     /// - The Pingora configuration fails.
     /// - The server cannot be started.
@@ -44,19 +48,19 @@ impl StartProxy {
         let mut server = Server::new(Some(Opt::default())).map_err(|err| err.to_string())?;
         server.bootstrap();
 
-        let proxy_config = super::ProxyConfig::load_config_from_file()?;
+        println!("Starting proxy with workers: {:?}", self.workers);
 
         let workers = self
             .workers
             .iter()
-            .map(|worker| Backend::new(worker).map_err(TxProverServiceError::BackendCreationFailed))
-            .collect::<Result<Vec<Backend>, TxProverServiceError>>()?;
+            .map(|worker| Backend::new(worker).map_err(ProvingServiceError::BackendCreationFailed))
+            .collect::<Result<Vec<Backend>, ProvingServiceError>>()?;
 
         if workers.is_empty() {
             warn!("Starting the proxy without any workers");
         }
 
-        let worker_lb = LoadBalancerState::new(workers, &proxy_config).await?;
+        let worker_lb = LoadBalancerState::new(workers, &self.proxy_config).await?;
 
         let health_check_service = background_service("health_check", worker_lb);
 
@@ -67,18 +71,17 @@ impl StartProxy {
         let mut update_workers_service =
             Service::new("update_workers".to_string(), updater_service);
         update_workers_service.add_tcp(
-            format!("{}:{}", proxy_config.host, proxy_config.workers_update_port).as_str(),
+            format!("{}:{}", self.proxy_config.host.clone(), self.proxy_config.workers_update_port)
+                .as_str(),
         );
 
         // Set up the load balancer
         let mut lb = http_proxy_service(&server.configuration, LoadBalancer(worker_lb));
 
-        let proxy_host = proxy_config.host;
-        let proxy_port = proxy_config.port.to_string();
-        lb.add_tcp(format!("{}:{}", proxy_host, proxy_port).as_str());
+        lb.add_tcp(format!("{}:{}", &self.proxy_config.host, self.proxy_config.port).as_str());
         let logic = lb
             .app_logic_mut()
-            .ok_or(TxProverServiceError::PingoraConfigFailed("app logic not found".to_string()))?;
+            .ok_or(ProvingServiceError::PingoraConfigFailed("app logic not found".to_string()))?;
         let mut http_server_options = HttpServerOptions::default();
 
         // Enable HTTP/2 for plaintext
@@ -89,7 +92,12 @@ impl StartProxy {
         let mut prometheus_service_http =
             pingora::services::listening::Service::prometheus_http_service();
         prometheus_service_http.add_tcp(
-            format!("{}:{}", proxy_config.prometheus_host, proxy_config.prometheus_port).as_str(),
+            format!(
+                "{}:{}",
+                self.proxy_config.metrics_config.prometheus_host,
+                self.proxy_config.metrics_config.prometheus_port
+            )
+            .as_str(),
         );
 
         server.add_service(prometheus_service_http);