+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [BREAKING] Renamed `AccountData` to `AccountFile` (#1116).
 - Implement transaction batch prover in Rust (#1112).
 - Added the `is_non_fungible_asset_issued` procedure to the `miden` library (#1125).
+- [BREAKING] Refactored config file for `miden-proving-service` to be based on environment variables (#1120).
 
 ## 0.7.2 (2025-01-28) - `miden-objects` crate only